+++ src/schemathesis/extra/pytest_plugin.py
@@ -1,25 +1,16 @@
-from typing import Any, Callable, Generator, List, Optional
+from typing import Any, Callable, Generator, List, Type
 
 import pytest
-from _pytest import nodes
+from _pytest import fixtures, nodes
 from _pytest.config import hookimpl
-from _pytest.python import Function, PyCollector  # type: ignore
+from _pytest.fixtures import FuncFixtureInfo
+from _pytest.python import Class, Function, FunctionDefinition, Metafunc, Module, PyCollector  # type: ignore
 from hypothesis.errors import InvalidArgument  # pylint: disable=ungrouped-imports
 
 from .._hypothesis import create_test
 from ..exceptions import InvalidSchema
 from ..models import Endpoint
 from ..utils import is_schemathesis_test
-
-
-@hookimpl(hookwrapper=True)  # type:ignore # pragma: no mutate
-def pytest_pycollect_makeitem(collector: nodes.Collector, name: str, obj: Any) -> Optional["SchemathesisCase"]:
-    """Switch to a different collector if the test is parametrized marked by schemathesis."""
-    outcome = yield
-    if is_schemathesis_test(obj):
-        outcome.force_result(SchemathesisCase(obj, name, collector))
-    else:
-        outcome.get_result()
 
 
 class SchemathesisCase(PyCollector):
@@ -36,17 +27,58 @@ class SchemathesisCase(PyCollector):
 
         Could produce more than one test item if
         parametrization is applied via ``pytest.mark.parametrize`` or ``pytest_generate_tests``.
+
+        This implementation is based on the original one in pytest, but with slight adjustments
+        to produce tests out of hypothesis ones.
         """
+        name = self._get_test_name(endpoint)
+        funcobj = self._make_test(endpoint)
+
+        cls = self._get_class_parent()
+        definition = FunctionDefinition(name=self.name, parent=self.parent, callobj=funcobj)
+        fixturemanager = self.session._fixturemanager
+        fixtureinfo = fixturemanager.getfixtureinfo(definition, funcobj, cls)
+
+        metafunc = self._parametrize(cls, definition, fixtureinfo)
+
+        if not metafunc._calls:
+            yield Function(name, parent=self.parent, callobj=funcobj, fixtureinfo=fixtureinfo)
+        else:
+            fixtures.add_funcarg_pseudo_fixture_def(self.parent, metafunc, fixturemanager)
+            fixtureinfo.prune_dependency_tree()
+
+            for callspec in metafunc._calls:
+                subname = "{}[{}]".format(name, callspec.id)
+                yield Function(
+                    name=subname,
+                    parent=self.parent,
+                    callspec=callspec,
+                    callobj=funcobj,
+                    fixtureinfo=fixtureinfo,
+                    keywords={callspec.id: True},
+                    originalname=name,
+                )
+
+    def _get_class_parent(self) -> Type:
+        clscol = self.getparent(Class)
+        return clscol.obj if clscol else None
+
+    def _parametrize(self, cls: Type, definition: FunctionDefinition, fixtureinfo: FuncFixtureInfo) -> Metafunc:
+        module = self.getparent(Module).obj
+        metafunc = Metafunc(definition, fixtureinfo, self.config, cls=cls, module=module)
+        methods = []
+        if hasattr(module, "pytest_generate_tests"):
+            methods.append(module.pytest_generate_tests)
+        if hasattr(cls, "pytest_generate_tests"):
+            methods.append(cls().pytest_generate_tests)
+        self.ihook.pytest_generate_tests.call_extra(methods, {"metafunc": metafunc})
+        return metafunc
+
+    def _make_test(self, endpoint: Endpoint) -> Callable:
         try:
-            hypothesis_item = create_test(endpoint, self.test_function)
+            return create_test(endpoint, self.test_function)
         except InvalidSchema:
-            hypothesis_item = lambda: pytest.fail("Invalid schema for endpoint")
-        items = self.ihook.pytest_pycollect_makeitem(
-            collector=self.parent, name=self._get_test_name(endpoint), obj=hypothesis_item
-        )
-        for item in items:
-            item.obj = hypothesis_item
-            yield item
+            return lambda: pytest.fail("Invalid schema for endpoint")
 
     def collect(self) -> List[Function]:  # type: ignore
         """Generate different test items for all endpoints available in the given schema."""
@@ -56,6 +88,16 @@ class SchemathesisCase(PyCollector):
             ]
         except Exception:
             pytest.fail("Error during collection")
+
+
+@hookimpl(hookwrapper=True)  # type:ignore # pragma: no mutate
+def pytest_pycollect_makeitem(collector: nodes.Collector, name: str, obj: Any) -> None:
+    """Switch to a different collector if the test is parametrized marked by schemathesis."""
+    outcome = yield
+    if is_schemathesis_test(obj):
+        outcome.force_result(SchemathesisCase(obj, name, collector))
+    else:
+        outcome.get_result()
 
 
 @hookimpl(hookwrapper=True)  # pragma: no mutate