+++ docs/changelog.rst
@@ -12,6 +12,11 @@ Added
 - Display RNG seed in the CLI output to allow test reproducing. `#267`_
 - Allow to specify seed in CLI.
 
+Fixed
+~~~~~
+
+- Pytest fixture parametrization via ``pytest_generate_tests``. `#280`_
+
 `0.15.0`_ - 2019-11-15
 ----------------------
 
@@ -408,6 +413,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#280: https://github.com/kiwicom/schemathesis/issues/280
 .. _#272: https://github.com/kiwicom/schemathesis/issues/272
 .. _#270: https://github.com/kiwicom/schemathesis/issues/270
 .. _#268: https://github.com/kiwicom/schemathesis/issues/268